+++ lib/src/tabbed_view.dart
@@ -1,9 +1,5 @@
 import 'dart:math' as math;
-import 'dart:ui';
-
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 import 'package:tabbed_view/src/content_area.dart';
 import 'package:tabbed_view/src/tab_button.dart';
 import 'package:tabbed_view/src/tab_data.dart';