+++ lib/src/tabbed_view_menu_widget.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:tabbed_view/src/tabbed_view_data.dart';
 import 'package:tabbed_view/src/theme/menu_theme_data.dart';
 import 'package:tabbed_view/src/theme/tabbed_view_theme_data.dart';